Collision (sort of), PVectors, Title Screen, Health
Switched position and velocity variables to PVectors (easier, cleaner,
and more versatile)
fixed typo that broke collision against top of screen
added collision, sort of. It's broken, but a start? It's commented out
of for now.
Sheep vs Sheep collision has same problem, but more dramatic because
there are so many, so it's commented out for now.
Moved the lerp away command into the time freeze contingency
Added health variable and alive boolean to Player class.
Added variable for color just in case.